Replace support for --password
[git-svn](https://git-scm.com/docs/git-svn) does not support `--password` anymore.

This commit removes the command line option for svn2git since it ultimately passed its value to `svn-git.

If the SVN transport requires a password, svn-git will always prompt the user for it, so now svn2git will detect the presence of the SVN2GIT_PASSWORD environment variable and pass its value in reply to the prompt.